device_agent/printers/datecs.py
"""
ClovUp Device Agent — Datecs Fiscal Printer Driver.

Implements the BasePrinter interface for Datecs FP series printers.
Communication via serial port using Datecs FP protocol.
"""

import logging
import serial
from printers.base import BasePrinter
from config import settings

logger = logging.getLogger(__name__)


class DatecsPrinter(BasePrinter):
    """
    Driver for Datecs fiscal printers (DP-25, DP-55, FP-700, etc.).

    Uses serial (RS-232 / USB-serial) communication with the
    Datecs fiscal protocol.
    """

    def __init__(self):
        self.port = settings.printer_port
        self.baudrate = settings.printer_baudrate

    def _open_serial(self) -> serial.Serial:
        return serial.Serial(
            port=self.port,
            baudrate=self.baudrate,
            bytesize=serial.EIGHTBITS,
            parity=serial.PARITY_NONE,
            stopbits=serial.STOPBITS_ONE,
            timeout=5,
        )

    def _send_command(self, cmd: int, data: str = "") -> str:
        """
        Send a command to the Datecs printer and read the response.
        Protocol: <01><LEN><SEQ><CMD><DATA><05><BCC><03>
        """
        # Placeholder — real implementation requires building
        # the full Datecs binary protocol frame.
        logger.info("Datecs CMD %02X: %s", cmd, data)
        raise NotImplementedError(
            "Datecs protocol implementation requires device-specific "
            "binary frame construction. See Datecs FP protocol documentation."
        )

    def print_receipt(
        self,
        items: list[dict],
        payments: list[dict],
        cashier: str = "",
        is_storno: bool = False,
        storno_reason: str = "",
        original_receipt: str = "",
    ) -> str:
        # CMD 0x30 - Open fiscal receipt
        # CMD 0x31 - Sale line
        # CMD 0x35 - Payment
        # CMD 0x38 - Close receipt
        logger.info("Datecs: print_receipt (%d items, storno=%s)", len(items), is_storno)
        raise NotImplementedError("Datecs print_receipt not yet implemented.")

    def safe_in(self, amount: float) -> None:
        # CMD 0x46 with positive amount
        logger.info("Datecs: safe_in %.2f", amount)
        raise NotImplementedError("Datecs safe_in not yet implemented.")

    def safe_out(self, amount: float) -> None:
        # CMD 0x46 with negative amount
        logger.info("Datecs: safe_out %.2f", amount)
        raise NotImplementedError("Datecs safe_out not yet implemented.")

    def x_report(self) -> dict:
        # CMD 0x69
        logger.info("Datecs: X-report")
        raise NotImplementedError("Datecs x_report not yet implemented.")

    def z_report(self) -> dict:
        # CMD 0x6A
        logger.info("Datecs: Z-report")
        raise NotImplementedError("Datecs z_report not yet implemented.")

    def reprint(self, receipt_number: str) -> None:
        # CMD 0x6D
        logger.info("Datecs: reprint %s", receipt_number)
        raise NotImplementedError("Datecs reprint not yet implemented.")

    def get_status(self) -> dict:
        # CMD 0x4A
        logger.info("Datecs: get_status")
        return {"printer": "datecs", "port": self.port, "status": "stub"}
