device_agent/requirements.txt
# ClovUp Device Agent dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pyserial==3.5
python-escpos==4.0.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
apscheduler==3.10.4
